Move SignedDistanceToPoint tests to distance_to_point_callback_test
Key changes:
- Refactor utility for making fcl::CollisionObjectd for reuse across tests.
- Minimally reframe the tests to run directly on the callback instead of
  through the ProximityEngine gateway.
