+++ geometry/test/proximity_engine_test.cc
@@ -1369,765 +1369,6 @@ GTEST_TEST(SignedDistanceToPointBroadphaseTest, MultipleThreshold) {
   EXPECT_EQ(2, results.size());
 }
 
-// Test the narrow-phase part of ComputeSignedDistanceToPoint.
-
-// Parameter for the value-parameterized test fixture SignedDistanceToPointTest.
-struct SignedDistanceToPointTestData {
-  SignedDistanceToPointTestData(shared_ptr<Shape> geometry_in,
-                                const RigidTransformd& X_WG_in,
-                                const Vector3d& p_WQ_in,
-                                const SignedDistanceToPoint<double>& expect_in)
-      : geometry(geometry_in),
-        X_WG(X_WG_in),
-        p_WQ(p_WQ_in),
-        expected_result(expect_in) {}
-
-  static SignedDistanceToPointTestData Make(shared_ptr<Shape> shape,
-                                            const RigidTransformd& X_WG,
-                                            const Vector3d& p_WQ,
-                                            const Vector3d& p_GN,
-                                            double distance,
-                                            const Vector3d& grad_W) {
-    return SignedDistanceToPointTestData{
-        shape, X_WG, p_WQ,
-        SignedDistanceToPoint{GeometryId::get_new_id(), p_GN, distance,
-                              grad_W}};
-  }
-
-  // Google Test uses this operator to report the test data in the log file
-  // when a test fails.
-  friend std::ostream& operator<<(std::ostream& os,
-                                  const SignedDistanceToPointTestData& obj) {
-    fmt::print(os,
-               "{{\n"
-               "  geometry: (not printed)\n"
-               "  X_WG: (not printed)\n"
-               "  p_WQ: {}\n"
-               "  expected_result.p_GN: {}\n"
-               "  expected_result.distance: {}\n"
-               "  expected_result.grad_W: {}\n"
-               "}}",
-               fmt_eigen(obj.p_WQ.transpose()),
-               fmt_eigen(obj.expected_result.p_GN.transpose()),
-               obj.expected_result.distance,
-               fmt_eigen(obj.expected_result.grad_W.transpose()));
-    return os;
-  }
-
-  shared_ptr<Shape> geometry;
-  const RigidTransformd X_WG;
-  const Vector3d p_WQ;
-  const SignedDistanceToPoint<double> expected_result;
-};
-
-std::vector<SignedDistanceToPointTestData> GenDistanceTestDataSphere(
-    const RigidTransformd& X_WG = RigidTransformd::Identity()) {
-  // We use these identities
-  //           2^2   + 3^2    + 6^2   = 7^2         (1)
-  //           0.2^2 + 0.3^2  + 0.6^2 = 0.7^2       (2)
-  //           0.1^2 + 0.15^2 + 0.3^2 = 0.35^2      (3)
-  // to set up the radius of the sphere geometry G and the position of the
-  // query point Q in such a way that both the signed distance and the
-  // position of the nearest point N can be expressed with fixed-point numbers.
-  //
-  // We will set a query point Q both outside G and inside G in such a way
-  // that Q is collinear to the center Go of the sphere and the nearest point N.
-  //
-  //            Q
-  //           /
-  //     ooo  /
-  //  o      N
-  // o      Q  o
-  // o    Go   o
-  // o         o
-  //  o       o
-  //     ooo
-  //
-  // From (2), we will set the radius of G to 0.7, so later we will have
-  // p_GN at (0.2, 0.3, 0.6). From (1), we will set p_GQ to (2,3,6), so it
-  // will be outside G at the positive distance 7 - 0.7 = 6.3. From (3), we will
-  // set p_GQ to (0.1, 0.15, 0.3), so it will be inside G at the negative
-  // distance 0.35 - 0.7 = -0.35. Both positions of Q will have the same
-  // nearest point N and the gradient vector.
-  auto sphere = make_shared<Sphere>(0.7);
-  std::vector<SignedDistanceToPointTestData> test_data{
-      // p_GQ = (2,3,6) is outside G at the positive distance 6.3 = 7 - 0.7.
-      SignedDistanceToPointTestData::Make(
-          sphere, X_WG, X_WG * Vector3d{2, 3, 6}, Vector3d(0.2, 0.3, 0.6), 6.3,
-          X_WG.rotation() * Vector3d(2, 3, 6) / 7),
-      // p_GQ = (0.1,0.15,0.3) is inside G at the negative distance -0.35.
-      SignedDistanceToPointTestData::Make(
-          sphere, X_WG, X_WG * Vector3d{0.1, 0.15, 0.3},
-          Vector3d(0.2, 0.3, 0.6), -0.35,
-          X_WG.rotation() * Vector3d(2, 3, 6) / 7),
-      // Reports an arbitrary gradient vector (as defined in the
-      // QueryObject::ComputeSignedDistanceToPoint() documentation) at the
-      // center of the sphere.
-      SignedDistanceToPointTestData::Make(
-          sphere, X_WG, X_WG * Vector3d{0, 0, 0}, Vector3d(0.7, 0, 0), -0.7,
-          X_WG.rotation() * Vector3d(1, 0, 0))};
-  return test_data;
-}
-
-std::vector<SignedDistanceToPointTestData> GenDistTestTransformSphere() {
-  const RigidTransformd X_WG(RollPitchYawd(M_PI / 6, M_PI / 3, M_PI_2),
-                             Vector3d{10, 11, 12});
-  return GenDistanceTestDataSphere(X_WG);
-}
-
-// We declare this function here, so we can call it from
-// GenDistanceTestDataOutsideBox() below.
-std::vector<SignedDistanceToPointTestData> GenDistanceTestDataBoxBoundary(
-    const RigidTransformd& X_WG = RigidTransformd::Identity());
-
-// Generates test data for a query point Q outside a box geometry G nearest
-// to one of the 6 faces, the 12 edges, and the 8 vertices of the box.
-// First we call GenDistanceTestDataBoxBoundary() to generate test data for
-// query points on the box boundary. Then, we move the query point along the
-// gradient vector by a unit distance.  In each case, the nearest point to Q
-// on ∂G stays the same, the signed distance becomes +1, and the gradient
-// vector stays the same.
-std::vector<SignedDistanceToPointTestData> GenDistanceTestDataOutsideBox(
-    const RigidTransformd& X_WG = RigidTransformd::Identity()) {
-  const std::vector<SignedDistanceToPointTestData> test_data_box_boundary =
-      GenDistanceTestDataBoxBoundary(X_WG);
-  std::vector<SignedDistanceToPointTestData> test_data;
-  for (const auto& data : test_data_box_boundary) {
-    const shared_ptr<Shape> shape = data.geometry;
-    // We expect the shape to be a box.
-    DRAKE_DEMAND(dynamic_cast<Box*>(shape.get()) != nullptr);
-    // The gradient grad_W has unit length by construction.
-    const Vector3d p_WQ = data.p_WQ + data.expected_result.grad_W;
-    const GeometryId& id = data.expected_result.id_G;
-    const Vector3d& p_GN = data.expected_result.p_GN;
-    const double distance = 1;
-    const Vector3d& grad_W = data.expected_result.grad_W;
-    test_data.emplace_back(
-        shape, X_WG, p_WQ,
-        SignedDistanceToPoint<double>(id, p_GN, distance, grad_W));
-  }
-  return test_data;
-}
-
-std::vector<SignedDistanceToPointTestData> GenDistTestTransformOutsideBox() {
-  RigidTransformd X_WG(RollPitchYawd(M_PI / 3, M_PI / 6, M_PI_2),
-                       Vector3d{10, 11, 12});
-  return GenDistanceTestDataOutsideBox(X_WG);
-}
-
-// Generates test data for a query point Q on the boundary ∂G of a box
-// geometry G. Q can be at the 8 corners, at the midpoints of the 12 edges, or
-// at the centers of the 6 faces of G. The set of all 26 positions can be
-// expressed in G's frame as the following Cartesian product excluding the
-// origin (3x3x3-1 = 26) :
-//
-//     p_GQ ∈ {-h(x),0,+h(x)} x {-h(y),0,+h(y)} x {-h(z),0,+h(z)} - {(0,0,0)},
-//
-// where h(x), h(y), and h(z) are the half width, half depth, and half height
-// of G, respectively. We do not allow p_GQ=(0,0,0) because it is in the
-// interior of G. The number of zeroes in p_GQ corresponds to the location at
-// a corner (no zero, 3 non-zeroes), at the midpoint of an edge (1 zero,
-// 2 non-zeroes), or at the center of a face (2 zeroes, 1 non-zero).
-//     The positions above is parameterized by the sign vector s expressed in
-// G's frame as:
-//
-//     s_G = (sx,sy,sz) ∈ {-1,0,+1} x {-1,0,+1} x {-1,0,+1} - {(0,0,0)},
-//     p_GQ(s) = s_G ∘ h_G,
-//
-// where h_G = (h(x), h(y), h(z)), which is the vector from the origin Go to a
-// vertex of G expressed in G's frame. The operator ∘ is the entrywise product
-// (also known as Hadamard product): (a,b,c)∘(u,v,w) = (a*u, b*v, c*w).
-//     In each case, Q is also its own nearest point on ∂G, the signed distance
-// is always zero, and the gradient vector equals the normalized unit vector of
-// the vector s.
-std::vector<SignedDistanceToPointTestData> GenDistanceTestDataBoxBoundary(
-    const RigidTransformd& X_WG) {
-  auto box = make_shared<Box>(20.0, 30.0, 10.0);
-  const Vector3d h_G = box->size() / 2;
-  const double distance = 0;
-  std::vector<SignedDistanceToPointTestData> test_data;
-  for (const double sx : {-1, 0, 1}) {
-    for (const double sy : {-1, 0, 1}) {
-      for (const double sz : {-1, 0, 1}) {
-        // Skip the origin.
-        if (sx == 0 && sy == 0 && sz == 0) continue;
-        const Vector3d s_G(sx, sy, sz);
-        const Vector3d p_GQ = s_G.cwiseProduct(h_G);
-        const Vector3d p_WQ = X_WG * p_GQ;
-        // We create new id for each test case to help distinguish them.
-        const GeometryId id = GeometryId::get_new_id();
-        // Q is its own nearest point on ∂G.
-        const Vector3d& p_GN = p_GQ;
-        // Rotation matrix for transforming vector expression from G to world.
-        const RotationMatrixd& R_WG = X_WG.rotation();
-        const Vector3d grad_G = s_G.normalized();
-        const Vector3d grad_W = R_WG * grad_G;
-        test_data.emplace_back(
-            box, X_WG, p_WQ,
-            SignedDistanceToPoint<double>(id, p_GN, distance, grad_W));
-      }
-    }
-  }
-  return test_data;
-}
-
-std::vector<SignedDistanceToPointTestData> GenDistTestTransformBoxBoundary() {
-  RigidTransformd X_WG(RollPitchYawd(M_PI / 3, M_PI / 6, M_PI_2),
-                       Vector3d{10, 11, 12});
-  return GenDistanceTestDataBoxBoundary(X_WG);
-}
-
-// Generates test data for a query point Q inside a box geometry G with a
-// unique nearest point on the boundary ∂G. Unlike Q on ∂G or outside G that
-// has 26 cases each, we have only 6 cases of Q inside G. In each case, Q is
-// unambiguously closest to a single face.
-//     The position of Q is parameterized by a chosen distance d and the
-// outward unit normal vector s of the six faces of G. We calculate the
-// center C of the face and offset C by the distance d inwards into G:
-//
-//         d    = min {h(x), h(y), h(z)} / 2,
-//         s_G  ∈ {-x, +x, -y, +y, -z, +z}
-//         p_GC = s_G ∘ h_G
-//         p_GQ = p_GC - d * s_G,
-//
-// where h_G = (h(x), h(y), h(z)) is the vector of the half width, half depth,
-// and half height of G. It is the vector from the origin Go to a corner
-// of G expressed in G's frame. The operator ∘ is the entry-wise product:
-// (a,b,c)∘(u,v,w) = (a*u, b*v, c*w).
-//     The chosen d is small enough that Q at the inward normal offset from a
-// face center is still unambiguously closest to that face.
-//     In each case, the nearest point N on ∂G is C, the negative signed
-// distance is -d, and the gradient vector is the face normal vector s.
-std::vector<SignedDistanceToPointTestData> GenDistTestDataInsideBoxUnique(
-    const RigidTransformd& X_WG = RigidTransformd::Identity()) {
-  // Create a box [-10,10]x[-15,15]x[-5,5],
-  auto box = make_shared<Box>(20.0, 30.0, 10.0);
-  const Vector3d h_G = box->size() / 2;
-  const double d = h_G.minCoeff() / 2;
-  std::vector<SignedDistanceToPointTestData> test_data;
-  for (const Vector3d unit_vector :
-       {Vector3d::UnitX(), Vector3d::UnitY(), Vector3d::UnitZ()}) {
-    for (const double sign : {-1, 1}) {
-      // Unit face normal vector.
-      const Vector3d s_G = sign * unit_vector;
-      // Center of a face.
-      const Vector3d p_GC = s_G.cwiseProduct(h_G);
-      const Vector3d p_GQ = p_GC - d * s_G;
-      const Vector3d p_WQ = X_WG * p_GQ;
-      // We create new id for each test case to help distinguish them.
-      const GeometryId id = GeometryId::get_new_id();
-      // The nearest point is at the face center.
-      const Vector3d& p_GN = p_GC;
-      // Rotation matrix for transforming vector expression from G to world.
-      const RotationMatrixd R_WG = X_WG.rotation();
-      const Vector3d& grad_G = s_G;
-      const Vector3d grad_W = R_WG * grad_G;
-      test_data.emplace_back(
-          box, X_WG, p_WQ, SignedDistanceToPoint<double>(id, p_GN, -d, grad_W));
-    }
-  }
-  return test_data;
-}
-
-// We test a rigid transform with the signed distance to query point Q inside
-// a box B only when Q has a unique nearest point on the boundary ∂B.
-std::vector<SignedDistanceToPointTestData>
-GenDistTestTransformInsideBoxUnique() {
-  RigidTransformd X_WG(RollPitchYawd(M_PI / 3, M_PI / 6, M_PI_2),
-                       Vector3d{10, 11, 12});
-  return GenDistTestDataInsideBoxUnique(X_WG);
-}
-
-// A query point Q inside a box G with multiple nearest points on the
-// boundary ∂B needs a tie-breaking rule. However, a rigid transform can
-// contaminate the tie breaking due to rounding errors.  We test the tie
-// breaking with the identity transform only.
-std::vector<SignedDistanceToPointTestData> GenDistTestDataInsideBoxNonUnique() {
-  // We set up a 20x10x10 box [-10,10]x[-5,5]x[-5,5].  Having the same depth
-  // and height allows Q to have five nearest faces in the last case below.
-  auto box = make_shared<Box>(20.0, 10.0, 10.0);
-  const RigidTransformd& X_WG = RigidTransformd::Identity();
-  std::vector<SignedDistanceToPointTestData> test_data{
-      // Q is nearest to the face [-10,10]x[-5,5]x{5}.
-      SignedDistanceToPointTestData::Make(box, X_WG, X_WG * Vector3d{6, 1, 2},
-                                          Vector3d(6, 1, 5), -3.0,
-                                          X_WG.rotation() * Vector3d(0, 0, 1)),
-      // Q is nearest to two faces {10}x[-5,5]x[-5,5] and [-10,10]x[-5,5]x{5}.
-      SignedDistanceToPointTestData::Make(box, X_WG, X_WG * Vector3d{6, 0, 1},
-                                          Vector3d(10, 0, 1), -4.0,
-                                          X_WG.rotation() * Vector3d(1, 0, 0)),
-      // Q is nearest to three faces {10}x[-5,5]x[-5,5], [-10,10]x{5}x[-5,5],
-      // and [-10,10]x[-5,5]x{5}.
-      SignedDistanceToPointTestData::Make(box, X_WG, X_WG * Vector3d{6, 1, 1},
-                                          Vector3d(10, 1, 1), -4.0,
-                                          X_WG.rotation() * Vector3d(1, 0, 0)),
-      // Q at the center of the box is nearest to four faces
-      // [-10,10]x{5}x[-5,5], [-10,10]x{-5}x[-5,5], [-10,10]x[5,-5]x{5},
-      // and [-10,10]x[5,-5]x{-5}.
-      SignedDistanceToPointTestData::Make(box, X_WG, X_WG * Vector3d{0, 0, 0},
-                                          Vector3d(0, 5, 0), -5.0,
-                                          X_WG.rotation() * Vector3d(0, 1, 0)),
-      // Q is nearest to five faces {10}x[-5,5]x[-5,5], [-10,10]x{5}x[-5,5],
-      // [-10,10]x{-5}x[-5,5], [-10,10]x[5,-5]x{5}, and [-10,10]x[5,-5]x{-5}.
-      SignedDistanceToPointTestData::Make(box, X_WG, X_WG * Vector3d{5, 0, 0},
-                                          Vector3d(10, 0, 0), -5.0,
-                                          X_WG.rotation() * Vector3d(1, 0, 0))};
-  return test_data;
-}
-
-std::vector<SignedDistanceToPointTestData> GenDistanceTestDataTranslateBox() {
-  // We set up a 20x10x30 box G centered at the origin [-10,10]x[-5,5]x[-15,15].
-  auto box = make_shared<Box>(20.0, 10.0, 30.0);
-  std::vector<SignedDistanceToPointTestData> test_data{
-      // We translate G by (10,20,30) to [0,20]x[15,25]x[15,45] in World frame.
-      // The position of the query point p_WQ(23,20,30) is closest to G at
-      // (20,20,30) in World frame, which is p_GN=(10,0,0) in G's frame.
-      // The gradient vector is expressed in both frames as (1,0,0).
-      SignedDistanceToPointTestData::Make(
-          box, Translation3d(10, 20, 30), Vector3d{23, 20, 30},
-          Vector3d(10, 0, 0), 3.0, Vector3d(1, 0, 0))};
-  return test_data;
-}
-
-// Generate test data for a query point Q from a cylinder G.
-
-// We separate the test data for Q on the boundary ∂G into two parts: Q on
-// the top/bottom circles (GenDistTestDataCylinderBoundaryCircle) and Q on
-// the cap/barrel surfaces (GenDistTestDataCylinderBoundarySurface).
-// Here, a circle is a 1-dimensional closed curve, and a cap is a
-// 2-dimensional flat surface bounded by a circle.
-//     We separate the two cases because we will generate Q outside/inside G
-// by moving Q on ∂G outwards/inwards along its gradient vector. In most
-// cases, we can move Q a small distance, and its nearest point N on ∂G stays
-// the same, namely the original Q on ∂G. However, moving Q on the top/bottom
-// circles inwards would change its nearest point N, and we don't want to do
-// that.
-//     Later we will combine them into GenDistTestDataCylinderBoundary which
-// will be used by GenDistTestDataOutsideCylinder.
-//     In summary, the call graph looks like this:
-//
-// GenDistTestDataInsideCylinder
-// |
-// |  GenDistTestDataOutsideCylinder
-// |  |
-// |  +--> GenDistTestDataCylinderBoundary
-// |       |
-// |       +--> GenDistTestDataCylinderBoundaryCircle
-// |       |
-// +-----> +--> GenDistTestDataCylinderBoundarySurface
-
-// Generates test data for a query point Q on the top/bottom circles of a
-// cylinder geometry G.
-std::vector<SignedDistanceToPointTestData>
-GenDistTestDataCylinderBoundaryCircle(
-    const RigidTransformd& X_WG = RigidTransformd::Identity()) {
-  const RotationMatrixd& R_WG = X_WG.rotation();
-  auto cylinder = make_shared<Cylinder>(3.0, 5.0);
-  const double radius = cylinder->radius();
-  const double half_length = cylinder->length() / 2;
-  // We want the test to cover all the combinations of positive, negative,
-  // and zero values of both x and y coordinates on the two boundary circles
-  // of the cylinder. Furthermore, each (x,y) has |x| ≠ |y| to avoid
-  // symmetry that might hide problems. We achieve this by having 12 points
-  // equally spread around a unit circle and map them to the two boundary
-  // circles later.
-  const int kNumVectors = 12;
-  // unit vectors in x-y plane of G's frame.
-  std::vector<Vector3d> all_xy_vectors;
-  const Vector3d kXVector(1, 0, 0);
-  for (int c = 0; c < kNumVectors; ++c) {
-    all_xy_vectors.push_back(
-        RotationMatrixd(RollPitchYawd(0, 0, 2 * M_PI * c / kNumVectors)) *
-        kXVector);
-  }
-  const double distance = 0;  // Q on ∂G has distance zero.
-  std::vector<SignedDistanceToPointTestData> test_data;
-  for (const auto& z_vector : {Vector3d(0, 0, 1), Vector3d(0, 0, -1)}) {
-    for (const auto& xy_vector : all_xy_vectors) {
-      const GeometryId id = GeometryId::get_new_id();
-      // xy_vector and z_vector are unit vectors in x-y plane and z-axis of G.
-      const Vector3d p_GQ = radius * xy_vector + half_length * z_vector;
-      const Vector3d p_WQ = X_WG * p_GQ;
-      // Q is its own nearest point on ∂G.
-      const Vector3d p_GN = p_GQ;
-      // We set the gradient vector according to the convention described in
-      // QueryObject::ComputeSignedDistanceToPoint().  Mathematically it is
-      // undefined.
-      const Vector3d grad_G = (xy_vector + z_vector).normalized();
-      const Vector3d grad_W = R_WG * grad_G;
-      test_data.emplace_back(
-          cylinder, X_WG, p_WQ,
-          SignedDistanceToPoint<double>(id, p_GN, distance, grad_W));
-    }
-  }
-  return test_data;
-}
-
-// Generates test data for a query point Q on the boundary surface ∂G of a
-// cylinder geometry G. Q can be on the barrel or on the top or bottom caps.
-std::vector<SignedDistanceToPointTestData>
-GenDistTestDataCylinderBoundarySurface(
-    const RigidTransformd& X_WG = RigidTransformd::Identity()) {
-  const RotationMatrixd& R_WG = X_WG.rotation();
-  auto cylinder = make_shared<Cylinder>(3.0, 5.0);
-  const double radius = cylinder->radius();
-  const double half_length = cylinder->length() / 2;
-  // We want the test to cover all the combinations of positive, negative,
-  // and zero values of both x and y coordinates on some circles on the
-  // barrel or on the caps. Furthermore, each (x,y) has |x| ≠ |y| to avoid
-  // symmetry that might hide problems. We achieve this goal by having
-  // 12 points equally spread around a unit circle and map them to the barrel
-  // or the caps later.
-  const int kNumVectors = 12;
-  // unit vectors in x-y plane of G's frame.
-  std::vector<Vector3d> all_xy_vectors;
-  const Vector3d kXVector(1, 0, 0);
-  for (int c = 0; c < kNumVectors; ++c) {
-    all_xy_vectors.push_back(
-        RotationMatrixd(RollPitchYawd(0, 0, 2 * M_PI * c / kNumVectors)) *
-        kXVector);
-  }
-  struct LocalTestData {
-    LocalTestData(const Vector3d& p_GQ_in, const Vector3d& grad_G_in)
-        : p_GQ(p_GQ_in), grad_G(grad_G_in) {}
-    Vector3d p_GQ;
-    Vector3d grad_G;
-  };
-  // Generate LocalTestData that will convert to
-  // SignedDistanceToPointTestData later.
-  std::vector<LocalTestData> test_data_barrel;
-  std::vector<LocalTestData> test_data_caps;
-  std::vector<LocalTestData> test_data_cap_centers;
-  for (const auto& z_vector : {Vector3d(0, 0, 1), Vector3d(0, 0, -1)}) {
-    // Q at the centers of the top and bottom caps.
-    {
-      // z_vector is a unit vector.
-      const Vector3d p_GQ = half_length * z_vector;
-      // The gradient vector on the circular disks is along the z-axis of G.
-      const Vector3d grad_G = z_vector;
-      test_data_cap_centers.emplace_back(p_GQ, grad_G);
-    }
-    for (const auto& xy_vector : all_xy_vectors) {
-      // Q on the barrel.
-      {
-        // Control how far vertically Q is from x-y plane of G.
-        const double kZFactor = 0.5;
-        // xy_vector and z_vector are unit vectors in x-y plane and z-axis of G.
-        const Vector3d p_GQ =
-            radius * xy_vector + half_length * kZFactor * z_vector;
-        // The gradient vector on the barrel is parallel to the x-y plane of G.
-        const Vector3d grad_G = xy_vector;
-        test_data_barrel.emplace_back(p_GQ, grad_G);
-      }
-      // Q on the top and bottom caps.
-      {
-        // Control how far radially Q is from z-axis of G.
-        const double kRFactor = 0.6;
-        // xy_vector and z_vector are unit vectors in x-y plane and z-axis of G.
-        const Vector3d p_GQ =
-            radius * kRFactor * xy_vector + half_length * z_vector;
-        // The gradient vector on the caps is along the z-axis of G.
-        const Vector3d grad_G = z_vector;
-        test_data_caps.emplace_back(p_GQ, grad_G);
-      }
-    }
-  }
-  std::vector<SignedDistanceToPointTestData> test_data;
-  // Convert LocalTestData to SignedDistanceToPointTestData and add to
-  // test_data.
-  auto convert = [&cylinder, &X_WG, &R_WG,
-                  &test_data](const LocalTestData& local) {
-    const Vector3d p_WQ = X_WG * local.p_GQ;
-    // Q on ∂G has distance zero.
-    const double distance = 0;
-    // Q is its own nearest point on ∂G.
-    const Vector3d p_GN = local.p_GQ;
-    const Vector3d grad_W = R_WG * local.grad_G;
-    // Implicitly generate a new geometry id for every test record.
-    test_data.push_back(SignedDistanceToPointTestData::Make(
-        cylinder, X_WG, p_WQ, p_GN, distance, grad_W));
-  };
-  std::for_each(test_data_barrel.begin(), test_data_barrel.end(), convert);
-  std::for_each(test_data_caps.begin(), test_data_caps.end(), convert);
-  std::for_each(test_data_cap_centers.begin(), test_data_cap_centers.end(),
-                convert);
-  return test_data;
-}
-
-// Generates test data for a query point Q on the boundary of a cylinder
-// geometry G. Combine GenDistTestDataCylinderBoundaryCircle with
-// GenDistTestDataCylinderBoundarySurface.
-std::vector<SignedDistanceToPointTestData> GenDistTestDataCylinderBoundary(
-    const RigidTransformd& X_WG = RigidTransformd::Identity()) {
-  auto test_data = GenDistTestDataCylinderBoundaryCircle(X_WG);
-  const auto test_data_boundary_surface =
-      GenDistTestDataCylinderBoundarySurface(X_WG);
-  for (const auto& data : test_data_boundary_surface) {
-    test_data.emplace_back(data.geometry, data.X_WG, data.p_WQ,
-                           data.expected_result);
-  }
-  return test_data;
-}
-
-// Generates test data for a query point Q outside a cylinder geometry G.
-// First we call GenDistTestDataCylinderBoundary() to generate test data for
-// query points on the boundary.  Then, we move the query point along the
-// gradient vector by a unit distance outward. The nearest point to Q on ∂G
-// stays the same, the signed distance becomes +1, and the gradient vector
-// stays the same.
-std::vector<SignedDistanceToPointTestData> GenDistTestDataOutsideCylinder(
-    const RigidTransformd& X_WG = RigidTransformd::Identity()) {
-  const auto test_data_cylinder_boundary =
-      GenDistTestDataCylinderBoundary(X_WG);
-  std::vector<SignedDistanceToPointTestData> test_data;
-  for (const auto& data : test_data_cylinder_boundary) {
-    const shared_ptr<Shape>& shape = data.geometry;
-    // We expect the shape to be a cylinder.
-    DRAKE_DEMAND(dynamic_cast<Cylinder*>(shape.get()) != nullptr);
-    // The gradient grad_W has unit length by construction.
-    const Vector3d p_WQ = data.p_WQ + data.expected_result.grad_W;
-    const GeometryId& id = data.expected_result.id_G;
-    const Vector3d& p_GN = data.expected_result.p_GN;
-    const double distance = 1;
-    const Vector3d& grad_W = data.expected_result.grad_W;
-    test_data.emplace_back(
-        shape, X_WG, p_WQ,
-        SignedDistanceToPoint<double>(id, p_GN, distance, grad_W));
-  }
-  return test_data;
-}
-
-// Generates test data for a query point Q inside a cylinder geometry G with
-// a unique nearest point on the boundary ∂G. Unlike Q on ∂G or outside G, Q
-// inside G cannot have its nearest point N on the top and bottom boundary
-// circles; however, it can have N on the top and bottom cap surfaces.
-// First we call GenDistTestDataCylinderBoundarySurface() to generate data on
-// the boundary surface.  Then, we move the query point along the gradient
-// vector a small negative distance into the interior.  The nearest point N
-// stays the same, the signed distance becomes the negative distance, and the
-// gradient vector stays the same.
-std::vector<SignedDistanceToPointTestData> GenDistTestDataInsideCylinder(
-    const RigidTransformd& X_WG = RigidTransformd::Identity()) {
-  const auto test_data_cylinder_boundary_surface =
-      GenDistTestDataCylinderBoundarySurface(X_WG);
-  std::vector<SignedDistanceToPointTestData> test_data;
-  const double kNegativeDistance = -0.1;
-  for (const auto& data : test_data_cylinder_boundary_surface) {
-    const shared_ptr<Shape>& shape = data.geometry;
-    // We expect the shape to be a cylinder.
-    DRAKE_DEMAND(dynamic_cast<Cylinder*>(shape.get()) != nullptr);
-    // The gradient grad_W has unit length by construction.
-    const Vector3d p_WQ =
-        data.p_WQ + kNegativeDistance * data.expected_result.grad_W;
-    const GeometryId& id = data.expected_result.id_G;
-    const Vector3d& p_GN = data.expected_result.p_GN;
-    const Vector3d& grad_W = data.expected_result.grad_W;
-    test_data.emplace_back(
-        shape, X_WG, p_WQ,
-        SignedDistanceToPoint<double>(id, p_GN, kNegativeDistance, grad_W));
-  }
-  return test_data;
-}
-
-// Generates test data for a query point Q at the center of a long cylinder
-// geometry G. Q's nearest point on ∂G is not unique. Our code picks the one
-// on the x's axis of G's frame.
-std::vector<SignedDistanceToPointTestData> GenDistTestDataCylinderCenter(
-    const RigidTransformd& X_WG = RigidTransformd::Identity()) {
-  const RotationMatrixd& R_WG = X_WG.rotation();
-  auto long_cylinder = make_shared<Cylinder>(1.0, 20.0);
-  const double radius = long_cylinder->radius();
-  const GeometryId id = GeometryId::get_new_id();
-  // The query point Q is at the center of the cylinder.
-  const Vector3d p_GQ(0, 0, 0);
-  const Vector3d p_WQ = X_WG * p_GQ;
-  const double distance = -radius;
-  // The nearest point N is on the x's axis of G's frame by convention.
-  const Vector3d p_GN = radius * Vector3d(1, 0, 0);
-  const Vector3d grad_G = Vector3d(1, 0, 0);
-  const Vector3d grad_W = R_WG * grad_G;
-  std::vector<SignedDistanceToPointTestData> test_data;
-  test_data.emplace_back(
-      long_cylinder, X_WG, p_WQ,
-      SignedDistanceToPoint<double>(id, p_GN, distance, grad_W));
-  return test_data;
-}
-
-// Generates test data for a query point Q on the boundary, inside, and
-// outside a cylinder with a rigid transform.
-std::vector<SignedDistanceToPointTestData> GenDistTestDataCylinderTransform() {
-  RigidTransformd X_WG(RollPitchYawd(M_PI / 3, M_PI / 6, M_PI_2),
-                       Vector3d{10, 11, 12});
-  std::vector<SignedDistanceToPointTestData> test_data;
-  for (const auto& record : GenDistTestDataCylinderBoundary(X_WG))
-    test_data.emplace_back(record);
-  for (const auto& record : GenDistTestDataOutsideCylinder(X_WG))
-    test_data.emplace_back(record);
-  for (const auto& record : GenDistTestDataInsideCylinder(X_WG))
-    test_data.emplace_back(record);
-  for (const auto& record : GenDistTestDataCylinderCenter(X_WG))
-    test_data.emplace_back(record);
-  return test_data;
-}
-
-// Generate test data for a query point Q relative to a half space.
-std::vector<SignedDistanceToPointTestData> GenDistTestDataHalfSpace() {
-  std::vector<SignedDistanceToPointTestData> test_data;
-
-  const RigidTransformd X_WG1(RollPitchYawd(M_PI / 3, M_PI / 6, M_PI_2),
-                              Vector3d{10, 11, 12});
-  auto hs = make_shared<HalfSpace>();
-  const GeometryId id = GeometryId::get_new_id();
-
-  for (const auto& X_WG : {RigidTransformd(), X_WG1}) {
-    const auto& R_WG = X_WG.rotation();
-    const Vector3d grad_W = R_WG.col(2);
-    const Vector3d p_GN(1.25, 1.5, 0);
-    const Vector3d p_WN = X_WG * p_GN;
-
-    // Outside the half space.
-    const double distance = 1.5;
-    const Vector3d p_WQ1 = p_WN + distance * grad_W;
-    test_data.emplace_back(
-        hs, X_WG, p_WQ1,
-        SignedDistanceToPoint<double>(id, p_GN, distance, grad_W));
-
-    // On the half space boundary.
-    const Vector3d p_WQ2 = p_WN;
-    test_data.emplace_back(
-        hs, X_WG, p_WQ2, SignedDistanceToPoint<double>(id, p_GN, 0.0, grad_W));
-
-    // Inside the half space.
-    const Vector3d p_WQ3 = p_WN - distance * grad_W;
-    test_data.emplace_back(
-        hs, X_WG, p_WQ3,
-        SignedDistanceToPoint<double>(id, p_GN, -distance, grad_W));
-  }
-
-  return test_data;
-}
-
-// This test fixture takes data generated by GenDistanceTestData*(),
-// GenDistTestData*(), and GenDistTestTransform*() above.
-struct SignedDistanceToPointTest
-    : public testing::TestWithParam<SignedDistanceToPointTestData> {
-  ProximityEngine<double> engine;
-  unordered_map<GeometryId, RigidTransformd> X_WGs;
-
-  // The tolerance value for determining equivalency between expected and
-  // tested results. The underlying algorithms have an empirically-determined,
-  // hard-coded tolerance of 1e-14 to account for loss of precision due to
-  // rigid transformations and this tolerance reflects that.
-  static constexpr double kTolerance = 1e-14;
-
-  SignedDistanceToPointTest() {
-    const auto& data = GetParam();
-    const GeometryId id = data.expected_result.id_G;
-    engine.AddAnchoredGeometry(*data.geometry, data.X_WG, id);
-    X_WGs[id] = data.X_WG;
-  }
-};
-
-TEST_P(SignedDistanceToPointTest, SingleQueryPoint) {
-  const auto& data = GetParam();
-
-  auto results = engine.ComputeSignedDistanceToPoint(data.p_WQ, X_WGs);
-  EXPECT_EQ(results.size(), 1);
-  EXPECT_EQ(results[0].id_G, data.expected_result.id_G);
-  EXPECT_TRUE(
-      CompareMatrices(results[0].p_GN, data.expected_result.p_GN, kTolerance))
-      << "Incorrect nearest point.";
-  EXPECT_NEAR(results[0].distance, data.expected_result.distance, kTolerance)
-      << "Incorrect signed distance.";
-  EXPECT_TRUE(CompareMatrices(results[0].grad_W, data.expected_result.grad_W,
-                              kTolerance))
-      << "Incorrect gradient vector.";
-}
-
-TEST_P(SignedDistanceToPointTest, SingleQueryPointWithThreshold) {
-  const auto& data = GetParam();
-
-  const double large_threshold = data.expected_result.distance + 0.01;
-  auto results =
-      engine.ComputeSignedDistanceToPoint(data.p_WQ, X_WGs, large_threshold);
-  // The large threshold allows one object in the results.
-  EXPECT_EQ(results.size(), 1);
-
-  const double small_threshold = data.expected_result.distance - 0.01;
-  results =
-      engine.ComputeSignedDistanceToPoint(data.p_WQ, X_WGs, small_threshold);
-  // The small threshold skips all objects.
-  EXPECT_EQ(results.size(), 0);
-}
-
-TEST_P(SignedDistanceToPointTest, SingleQueryPointSymbolic) {
-  const auto& data = GetParam();
-  const double large_threshold = data.expected_result.distance + 0.01;
-
-  std::unique_ptr<ProximityEngine<Expression>> sym_engine =
-      engine.ToScalarType<Expression>();
-  auto sym_results = sym_engine->ComputeSignedDistanceToPoint(
-      data.p_WQ.cast<Expression>(),
-      {{data.expected_result.id_G, data.X_WG.cast<Expression>()}},
-      large_threshold);
-  // No geometries are supported yet for Expression. Currently, this call
-  // succeeds, but will always return empty results.
-  EXPECT_EQ(sym_results.size(), 0);
-}
-
-// To debug a specific test, you can use Bazel flag --test_filter and
-// --test_output.  For example, you can use the command:
-// ```
-//   bazel test //geometry:proximity_engine_test
-//       --test_filter=Sphere/SignedDistanceToPointTest.SingleQueryPoint/0
-//       --test_output=all
-// ```
-// to run the first case from the test data generated by
-// GenDistanceTestDataSphere() with the function
-// TEST_P(SignedDistanceToPointTest, SingleQueryPoint).
-// Sphere
-INSTANTIATE_TEST_SUITE_P(Sphere, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistanceTestDataSphere()));
-
-INSTANTIATE_TEST_SUITE_P(TransformSphere, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistTestTransformSphere()));
-
-// Box
-INSTANTIATE_TEST_SUITE_P(OutsideBox, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistanceTestDataOutsideBox()));
-INSTANTIATE_TEST_SUITE_P(BoxBoundary, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistanceTestDataBoxBoundary()));
-INSTANTIATE_TEST_SUITE_P(InsideBoxUnique, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistTestDataInsideBoxUnique()));
-INSTANTIATE_TEST_SUITE_P(
-    InsideBoxNonUnique, SignedDistanceToPointTest,
-    testing::ValuesIn(GenDistTestDataInsideBoxNonUnique()));
-INSTANTIATE_TEST_SUITE_P(TranslateBox, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistanceTestDataTranslateBox()));
-INSTANTIATE_TEST_SUITE_P(TransformOutsideBox, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistTestTransformOutsideBox()));
-INSTANTIATE_TEST_SUITE_P(TransformBoxBoundary, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistTestTransformBoxBoundary()));
-INSTANTIATE_TEST_SUITE_P(
-    TransformInsideBoxUnique, SignedDistanceToPointTest,
-    testing::ValuesIn(GenDistTestTransformInsideBoxUnique()));
-
-// Cylinder
-INSTANTIATE_TEST_SUITE_P(CylinderBoundary, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistTestDataCylinderBoundary()));
-INSTANTIATE_TEST_SUITE_P(OutsideCylinder, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistTestDataOutsideCylinder()));
-INSTANTIATE_TEST_SUITE_P(InsideCylinder, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistTestDataInsideCylinder()));
-INSTANTIATE_TEST_SUITE_P(CenterCylinder, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistTestDataCylinderCenter()));
-INSTANTIATE_TEST_SUITE_P(CylinderTransform, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistTestDataCylinderTransform()));
-
-// Half space
-INSTANTIATE_TEST_SUITE_P(Halfspace, SignedDistanceToPointTest,
-                         testing::ValuesIn(GenDistTestDataHalfSpace()));
-
 // Penetration tests -- testing data flow; not testing the value of the query.
 
 // A scene with no geometry reports no penetrations.
